+++ HISTORY.md
@@ -3,6 +3,7 @@ Dev version
 * Bugfix : invalid negative socket id 
 * Add optional TCP_NODELAY support
 * Add worker thread weight
+* Add skynet.queue
 
 v0.4.1 (2014-7-7)
 -----------